feat: cria componentes ProductGallery e ProductInfo
